requirements.txt
# Web / API (Async-First)
fastapi
uvicorn
flask-cors

# LLM & RAG Engine
openai
faiss-cpu
torch
transformers
numpy
tqdm
flask-limiter
